# Move tag query evaluation logic into tag expressions