--- a/api/cluster.go
+++ b/api/cluster.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"regexp"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 
@@ -264,12 +265,16 @@ func (s *Server) indexTagDelSeries(ctx *middleware.Context, request models.Index
 		}
 
 		expressions := make(tagquery.Expressions, len(tags))
+		builder := strings.Builder{}
 		for i := range tags {
-			expressions[i] = tagquery.Expression{
-				Tag:                   tags[i],
-				Operator:              tagquery.EQUAL,
-				RequiresNonEmptyValue: true,
+			tags[i].StringIntoBuilder(&builder)
+			var err error
+			expressions[i], err = tagquery.ParseExpression(builder.String())
+			if err != nil {
+				response.Write(ctx, response.WrapErrorForTagDB(err))
+				return
 			}
+			builder.Reset()
 		}
 
 		query, err := tagquery.NewQuery(expressions, 0)