--- a/api/cluster.go
+++ b/api/cluster.go
@@ -603,7 +603,7 @@ func (s *Server) indexMetaTagRecordUpsert(ctx *middleware.Context, req models.In
 
 	response.Write(ctx, response.NewMsgp(200, &models.MetaTagRecordUpsertResult{
 		MetaTags: result.MetaTags.Strings(),
-		Queries:  result.Queries.Strings(),
+		Queries:  result.Expressions.Strings(),
 		Created:  created,
 	}))
 }