--- a/api/cluster.go
+++ b/api/cluster.go
@@ -268,7 +268,7 @@ func (s *Server) indexTagDelSeries(ctx *middleware.Context, request models.Index
 		builder := strings.Builder{}
 		for i := range tags {
 			tags[i].StringIntoBuilder(&builder)
-			var err error
+
 			expressions[i], err = tagquery.ParseExpression(builder.String())
 			if err != nil {
 				response.Write(ctx, response.WrapErrorForTagDB(err))