--- a/docker/docker-cluster-query/metrictank.ini
+++ b/docker/docker-cluster-query/metrictank.ini
@@ -429,6 +429,11 @@ schema-file = /etc/metrictank/schema-idx-cassandra.toml
 # instruct the driver to not attempt to get host info from the system.peers table
 disable-initial-host-lookup = false
 
+### tag query evaluation
+[tag-query]
+# size of regular expression cache in tag query evaluation
+match-cache-size = 1000
+
 ### in-memory only
 [memory-idx]
 enabled = false
@@ -436,8 +441,6 @@ enabled = false
 tag-support = false
 # number of workers to spin up to evaluate tag queries
 tag-query-workers = 50
-# size of regular expression cache in tag query evaluation
-match-cache-size = 1000
 # path to index-rules.conf file
 rules-file = /etc/metrictank/index-rules.conf
 # maximum duration each second a prune job can lock the index.