Moves the tag query expression evaluation logic into the expression types. This provides us more flexibility to combine expressions from meta records and tag queries when we have to take both indexes into account.

One notable change to the logic is that MetricDefinition filters now assume that the tag index they're currently looking at isn't necessarily the only tag index. 
F.e. if an expression says `tag1!=abc` and we see that `metric1` does not have `tag1` defined via its metric (intrinsic) tags, then we can't be sure yet whether another index (f.e. meta tag index) does assign `tag1=abc` to `metric1`. So in this case the filter would return a `None` decision, which means that then other index(es) would need to get checked for whether this expression is satisfied or not. 
On the other hand if the metric (intrinsic) tags would assign `tag1=abc` to `metric1`, then it would directly return a conclusive decision `Fail` and sub-subsequent index(es) can be ignored. If the metric (intrinsic) tags would assign `tag1=otherValue` to `metric1`, then it would directly return a conclusive decision `Pass` because the indexes are checked in order of their priority of conflict resolution.

Currently the lookups in the meta tag index are not implemented yet, this will be coming in the next PR.

The current status of this PR is that it works fine in my manual tests and the unit tests pass. I'm going to add more benchmarks and potentially add some performance improvements, but I'm not expecting any major changes.

This implements step `2)` of this list: https://github.com/grafana/metrictank/issues/660#issuecomment-490490846

UPDATE, trying to explain the above stuff a bit better:

Filter Functions (Matchers):
The given query expressions, once parsed into the types implementing `tagquery.Expression`, now provide a method called `GetMetricDefinitionFilter()`. This method returns a filter function (in other contexts this is also called a "matcher") which can be passed a MetricDefinition and it will return a decision which is either `Pass`, `Fail` or `None` (inconclusive). This allows us to build a chain of filter functions, first from the given expressions, but also from the query expressions associated with meta records which match the given expressions. 
So let's say we have an expression `tag=~val.*` and want to apply it as a filter (not to build the initial result set). Once it's parsed into the according `tagquery.Expression` we call `GetMetricDefinitionFilter()` in `TagQueryContext.prepareExpressions()` to obtain a filter function which we'll use to filter down the result set. 
Here comes the important part -> 
Additionally, once we want to use the meta tag index, we can use the same `tagquery.Expression` instance to lookup meta tag records (`tagquery.MetaTagRecord`) which match the expression `tag=~val.*` from the meta tag index (`UnpartitionedMemoryIdx.metaTags`). When we do that lookup from the meta tag index we use the `tagquery.Expression` instance not as a filter, but to build a set of meta tag records which match the expression `tag=~val.*`. We do that via its methods `GetKey()` and `ValuePasses()` (because the `=~` operator matches against the tag value of a key).
The resulting meta tag records come with one or multiple sets of query expressions (`MetaTagRecord.Expressions`). If any of these sets of query expressions match a `MetricDefinition` this means that the meta tags associated with the meta tag record get assigned to that Metric and we already know that one of the meta tags matches the expression `tag=~val.*`, which means that the expression `tag=~val.*` matches this metric (via the meta tag that gets assigned to it). 
Concretely, to implement this, `prepareExpressions()` will lookup those meta tag records as described and build a set of filter chains from them (because there are multiple sets of expressions associated with each meta tag record, and one expression can result in multiple meta tag records). These filter chains can then be used by `testByAllExpressions()`. If a MetricDefinition passes any one of the meta tag record based filter chains then it should be part of the result set (so there's a logical OR there).
There will probably be some tuning regarding which filter chain is cheaper to evaluate and which is more expensive, we should try to filter the result set down using the cheap filters first before applying the expensive ones, that's why `tagquery.Expression` has a `GetCostMultiplier()` method, to estimate cost of evaluation.

Filter Decisions:
The metric definition filters (`tagquery.MetricDefinitionFilter`) return a filter decision (`tagquery.FilterDecision`) which can be one of `Fail`, `Pass` or `None`. The `Fail` and `Pass` value simply mean that a filter has been able to conclusively decide that a given `MetricDefinition` has either passed the filter or has been disqualified by it. The `None` case is necessary because in certain cases a filter cannot make this decision without also looking at the other indexes that need to be taken into account. Let's take the example `tag!=value`:
Once we use the meta tag index, the function `testByAllExpressions()` will first apply a filter function that's been generated from the expression `tag!=value` to each `MetricDefinition` of the result set. If a `MetricDefinition` comes with a tag `tag=value`, the filter can directly return `Fail`. If the `MetricDefinition` comes with a tag `tag=otherValue`, the filter can directly return `Pass`. But if the `MetricDefinition` does not have the tag `tag` defined at all, then this filter can't make a final decision because it is possible that one of the meta tag records assigns the tag `tag=value` to the given `MetricDefinition`, so it returns `None`. In the case of `None`, `testByAllExpressions()` will have to run the given `MetricDefinition` through the filter chains that have been generated from the meta tag records assigning the tag `tag=value` (as described above). If one of them matches the given `MetricDefinition`, then it can be disqualified at that point and we can interrupt the filter chains. If none of them match the given `MetricDefinition`, the `tagquery.Expression` object's `GetDefaultDecision()` is used to make the final decision whether the `MetricDefinition` matches this expression or not. In the case of the type `tagquery.expressionNotEqual` (representing `!=`) the return value of `GetDefaultDecision()` would be `Pass`. In the case of other expression types, like for example `tagquery.expressionEqual` (`=`) the return value of `GetDefaultDecision()` would be `Fail`. Because if a tag does not get assigned to a metric and we're evaluating a `=` expression, then the decision should be `Fail`. If a tag does not get assigned to a metric and we're evaluating a `!=` expression, then the result should be `Pass`.