@replay 2019-07-16
Just FYI, those are the current benchmark results compared between master and this branch:
```
replay@mst-nb:~$  benchcmp /tmp/bench_master  /tmp/bench_branch
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                11478558      40059410      +248.99%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              47054050      179546200     +281.57%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       31461560      44589916      +41.73%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     156438270     200434630     +28.12%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                15807         15468         -2.14%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              2261          2217          -1.95%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           19747003      9970316       -49.51%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         66596885      10764884      -83.84%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           9365065       9606763       +2.58%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         28489946      34461254      +20.96%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           49737850      42749714      -14.05%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         297312400     244258900     -17.84%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          62335045      57052060      -8.48%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        359227266     335862066     -6.50%

benchmark                                                        old allocs     new allocs     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                7690           5233           -31.95%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              3338           883            -73.55%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       235127         34635          -85.27%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     205174         4759           -97.68%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                32             32             +0.00%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              4              4              +0.00%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           26992          983            -96.36%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         3460           206            -94.05%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           17050          17009          -0.24%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         16615          16607          -0.05%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           370968         34930          -90.58%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         369369         33379          -90.96%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          501437         164938         -67.11%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        500794         164792         -67.09%

benchmark                                                        old bytes     new bytes     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                709176        511772        -27.84%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              310965        131580        -57.69%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       18413345      2381103       -87.07%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     16453477      419914        -97.45%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                2624          2624          +0.00%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              352           352           +0.00%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           1704966       211470        -87.60%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         240589        36368         -84.88%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           4691539       4687239       -0.09%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         4068719       4067700       -0.03%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           36166214      9278016       -74.35%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         35042364      8164630       -76.70%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          68448358      41549945      -39.30%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        63688786      36831418      -42.17%
```
Most of them are looking good, but two show a significant slow-down. I'll investigate where that slow down comes from and try to fix it.