@replay 2019-07-16
The good news is that I now have a good understand of why some benchmarks are performing worse, the bad news is that I have no good idea how to improve that. 

- In the current master we only have one tag index. This means that to evaluate the condition `tag=value` we only needs to check whether a certain id exists in the tag index under the keys `[tag][value]`. If it does the condition is satisfied, otherwise it is not.
- We now need to work based on the assumption that there are multiple indexes that can assign tags to a metric, with a defined order of precedence. This means when we see that a metric does not exist in `[tag][value]`, then we need to iterate over its tags and check if any of them assign the tag `tag` with a different value to it. If that's the case then we can directly return saying that the condition is not satisfied, because we check the different tag indexes in decreasing order of precedence. Only if by iterating over a metric's `.Tags` property we don't see the tag that we're looking for assigned with a different value, then we go on to check the next index.

The difference in these benchmarks comes from the additional looping over the `.Tags` property if there was no exact match for `tag=value`, to check whether there's another entry beginning with `tag=`.