@replay 2019-07-17
I think to solve this above issue with the performance regression I'll add another option similar to `tag-support = true/false`, I'll call it `meta-tag-support = true/false`. That way it should be possible to avoid performance regressions at least when meta tag support is turned off.