@replay 2019-07-18
With the latest 3 commits the performance has gotten better, but not as good as master yet:
```
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                40323084      31360866      -22.23%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              180773119     149006891     -17.57%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       45430404      45239445      -0.42%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     199925766     195143375     -2.39%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                17038         17804         +4.50%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              2291          2328          +1.62%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           9962729       9969944       +0.07%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         10184006      10596547      +4.05%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           11629725      11567934      -0.53%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         34105115      29431685      -13.70%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           47443844      47445716      +0.00%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         238119214     239374546     +0.53%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          72438808      74312474      +2.59%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        317500930     323824766     +1.99%

benchmark                                                        old allocs     new allocs     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                5159           5123           -0.70%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              825            822            -0.36%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       34615          34623          +0.02%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     4758           4759           +0.02%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                32             32             +0.00%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              4              4              +0.00%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           984            991            +0.71%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         206            207            +0.49%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           17009          17000          -0.05%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         16610          16611          +0.01%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           34920          34927          +0.02%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         33355          33357          +0.01%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          164903         164917         +0.01%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        164780         164776         -0.00%

benchmark                                                        old bytes     new bytes     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                496563        496810        +0.05%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              104305        101565        -2.63%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       2357742       2358837       +0.05%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     395076        388860        -1.57%
BenchmarkTagQueryKeysByPrefixSimple/partitioned-8                2624          2624          +0.00%
BenchmarkTagQueryKeysByPrefixSimple/unPartitioned-8              352           352           +0.00%
BenchmarkTagQueryKeysByPrefixExpressions/partitioned-8           210184        210344        +0.08%
BenchmarkTagQueryKeysByPrefixExpressions/unPartitioned-8         33915         33963         +0.14%
BenchmarkTagQueryFilterByEqualExpression/partitioned-8           4687368       4687182       -0.00%
BenchmarkTagQueryFilterByEqualExpression/unPartitioned-8         4068257       4068524       +0.01%
BenchmarkTagQueryFilterByMatchExpression/partitioned-8           9274441       9274677       +0.00%
BenchmarkTagQueryFilterByMatchExpression/unPartitioned-8         8120561       8120910       +0.00%
BenchmarkTagQueryFilterByHasTagExpression/partitioned-8          41543443      41544661      +0.00%
BenchmarkTagQueryFilterByHasTagExpression/unPartitioned-8        36818300      36818546      +0.00%
```