@replay 2019-07-18
With the latest push things have improved a lot. I discovered today that another reason why the benchmarks in this working branch were showing worse performance was because the sorting by expression cost in this branch was not as good anymore as in master. So I fixed that, now we have better sorting again and together with the other improvements that I added the benchmarks actually look really good now.

This is comparing master with the latest commit in this branch (`37b3.fa`), running those benchmarks which previously have shown a worse performance:

```
replay@mst-nb:~/go/src/github.com/grafana/metrictank/idx/memory$ benchcmp /tmp/bench.d4047399 /tmp/bench.37b3faf6
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                15076606      9971341       -33.86%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              49369377      11758999      -76.18%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       40507835      30111017      -25.67%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     145715884     156101314     +7.13%

benchmark                                                        old allocs     new allocs     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                7623           587            -92.30%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              3282           248            -92.44%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       235135         25169          -89.30%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     205170         4114           -97.99%

benchmark                                                        old bytes     new bytes     delta
BenchmarkTagQueryFilterAndIntersect/partitioned-8                699451        218019        -68.83%
BenchmarkTagQueryFilterAndIntersect/unPartitioned-8              298730        62456         -79.09%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/partitioned-8       18416190      1819865       -90.12%
BenchmarkTagQueryFilterAndIntersectOnlyRegex/unPartitioned-8     16430234      350143        -97.87%
```